requirements.txt
accelerate==1.10.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.12.1
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.64.0
anyio==4.11.0
argcomplete==3.6.2
asgiref==3.9.1
astor==0.8.1
asyncpg==0.30.0
attrs==25.3.0
audioop-lts==0.2.2
Authlib==1.6.5
babel==2.17.0
backoff==2.2.1
bandit==1.8.6
bcrypt==4.1.1
beautifulsoup4==4.13.4
bidict==0.23.1
billiard==3.6.4.0
black==23.11.0
blinker==1.9.0
boltons==21.0.0
boto3==1.40.4
botocore==1.40.4
bracex==2.6
Brotli==1.1.0
build==1.3.0
CacheControl==0.14.3
cachetools==5.5.2
celery==5.2.7
certifi==2025.6.15
cffi==1.17.1
channels==4.3.1
charset-normalizer==3.4.2
chromadb==1.0.15
click==8.1.8
click-didyoumean==0.3.1
click-option-group==0.5.9
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
ConfigArgParse==1.7.1
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.7
cron-descriptor==1.4.5
cryptography==43.0.1
datamodel-code-generator==0.33.0
datasets==4.0.0
Deprecated==1.2.18
dill==0.3.8
distro==1.9.0
dj-database-url==3.0.1
Django==4.2.17
django-appconf==1.1.0
django-celery-beat==2.8.1
django-compressor==4.5.1
django-cors-headers==4.7.0
django-crum==0.7.9
django-extensions==4.1
django-filter==25.1
django-otp==1.6.1
django-phonenumber-field==8.1.0
django-storages==1.14.6
django-timezone-field==7.1
django_celery_results==2.6.0
django_debug_toolbar==3.8.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
docstring_parser==0.17.0
dparse==0.6.4
drf-yasg==1.21.7
durationpy==0.10
ecdsa==0.19.1
einops==0.8.1
email-validator==2.1.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
execnet==2.1.1
face==24.0.0
factory-boy==3.3.0
Faker==19.2.0
fastapi==0.104.1
ffmpeg-python==0.2.0
filelock==3.16.1
filetype==1.2.0
firebase_admin==7.1.0
flake8==6.1.0
Flask==3.1.1
flask-babel==4.0.0
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.3.0
future==1.0.0
genson==1.3.0
gevent==25.5.1
geventhttpclient==2.3.4
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.45
glom==22.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.106.0
google-cloud-bigquery==3.35.1
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-resource-manager==1.14.2
google-cloud-storage==3.3.1
google-crc32c==1.7.1
google-genai==1.28.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
greenlet==3.2.3
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httplib2==0.30.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.3
huggingface-hub==0.34.3
humanfriendly==10.0
hyperframe==6.1.0
icalendar==6.3.1
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflect==7.5.0
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonfield==3.2.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kombu==5.5.4
kubernetes==33.1.0
langchain==0.3.27
langchain-core==0.3.72
langchain-google-genai==2.1.9
langchain-text-splitters==0.3.9
langsmith==0.4.11
locust==2.37.14
locust-cloud==1.26.3
loguru==0.7.3
lxml==6.0.0
Mako==1.3.10
markdown-it-py==3.0.0
markdown2==2.5.4
MarkupSafe==3.0.2
marshmallow==4.0.0
mccabe==0.7.0
mcp==1.16.0
mdurl==0.1.2
minify_html==0.16.4
mmh3==5.2.0
mongomock==4.3.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
mypy==1.7.1
mypy_extensions==1.1.0
mysqlclient==2.2.0
narwhals==2.0.1
networkx==3.5
nh3==0.3.0
nltk==3.9.1
numpy==2.3.1
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.6.1
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-django==0.57b0
opentelemetry-instrumentation-logging==0.57b0
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-instrumentation-wsgi==0.57b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
orjson==3.9.15
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.3.2
pandoc==2.4
passlib==1.7.4
pathspec==0.12.1
pbr==6.1.1
pdfminer.six==20250506
pdfplumber==0.11.7
peewee==3.18.3
phonenumbers==9.0.10
pillow==11.3.0
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.8
playwright==1.55.0
pluggy==1.6.0
plumbum==1.9.0
ply==3.11
posthog==5.4.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.33.1
psutil==6.1.1
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2-binary==2.9.10
pyarrow==21.0.0
pyasn1==0.4.8
pyasn1_modules==0.4.2
pybase64==1.4.2
pycodestyle==2.11.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pyee==13.0.0
pyflakes==3.1.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.14.0
PyMySQL==1.1.1
pyotp==2.9.0
pyparsing==3.2.3
pypdf==5.9.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
pytesseract==0.3.13
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-base-url==2.1.0
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-html==4.1.1
pytest-json-report==1.5.0
pytest-metadata==3.1.1
pytest-mock==3.14.1
pytest-playwright==0.7.1
pytest-xdist==3.8.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.0
python-engineio==4.12.2
python-jose==3.4.0
python-json-logger==3.3.0
python-magic==0.4.27
python-multipart==0.0.18
python-slugify==8.0.4
python-socketio==5.13.0
python-telegram-bot==22.3
pytz==2023.3
PyYAML==6.0.2
pyzmq==27.0.1
qrcode==8.2
rcssmin==1.1.2
redis==5.0.1
referencing==0.36.2
regex==2025.7.34
reportlab==4.4.4
requests==2.32.4
requests-mock==1.12.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.5.3
rjsmin==1.2.2
rpds-py==0.26.0
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
s3transfer==0.13.1
safetensors==0.5.3
safety==3.7.0
safety-schemas==0.0.16
scikit-learn==1.7.1
scipy==1.16.1
selenium==4.35.0
semgrep==1.144.0
sentence-transformers==5.1.0
sentinels==1.0.0
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.7
SpeechRecognition==3.14.3
SQLAlchemy==2.0.36
sqlparse==0.5.3
sse-starlette==3.0.3
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.27.0
stevedore==5.4.1
stream-zip==0.0.83
streamlit==1.48.0
sympy==1.14.0
tenacity==8.5.0
text-unidecode==1.3
threadpoolctl==3.6.0
thrift==0.22.0
tokenizers==0.21.4
toml==0.10.2
tomli==2.0.2
tomlkit==0.13.3
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.54.1
trio==0.30.0
trio-websocket==0.12.2
typeguard==4.4.4
typer==0.16.0
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
Unidecode==1.4.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcmatch==8.5.2
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
xlrd==2.0.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zope.event==5.1.1
zope.interface==7.2
zstandard==0.23.0
